refactor: 도메인 로직으로 empty validate 로직 위임
