+++ src/test/java/kitchenpos/domain/OrderTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain;
 
-import static kitchenpos.common.fixtures.OrderTableFixtures.ORDER_TABLE1;
+import static kitchenpos.common.fixtures.OrderTableFixtures.ORDER_TABLE1_NUMBER_OF_GUESTS;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import kitchenpos.exception.OrderException;
@@ -10,10 +10,22 @@ import org.junit.jupiter.api.Test;
 class OrderTest {
 
     @Test
+    @DisplayName("주문 생성 시 주문 테이블이 비어있는 상태면 예외가 발생한다")
+    void throws_OrderTableIsEmpty() {
+        // given
+        final OrderTable emptyOrderTable = new OrderTable(ORDER_TABLE1_NUMBER_OF_GUESTS, true);
+
+        // when & then
+        assertThatThrownBy(() -> Order.from(emptyOrderTable))
+                .isInstanceOf(OrderException.CannotOrderStateByOrderTableEmptyException.class)
+                .hasMessage("[ERROR] 주문 테이블이 비어있는 상태일 때 주문할 수 없습니다.");
+    }
+
+    @Test
     @DisplayName("OrderStatus 변경 시 현재 OrderStatus가 COMPLETION이면 Status를 변경할 수 없으므로 예외가 발생한다.")
     void validateAvailableChangeStatus() {
         // given
-        final Order order = Order.from(ORDER_TABLE1());
+        final Order order = Order.from(new OrderTable(ORDER_TABLE1_NUMBER_OF_GUESTS, false));
         order.changeStatus(OrderStatus.COMPLETION);
 
         // when & then
