+++ src/test/java/kitchenpos/domain/OrderValidatorTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.domain;
 
-import static kitchenpos.common.fixtures.OrderTableFixtures.ORDER_TABLE1_NUMBER_OF_GUESTS;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import kitchenpos.exception.OrderException;
@@ -20,17 +19,5 @@ class OrderValidatorTest {
         assertThatThrownBy(() -> OrderValidator.validateOrderLineItemSize(orderLineItemSize, foundMenuSize))
                 .isInstanceOf(OrderException.NotFoundOrderLineItemMenuExistException.class)
                 .hasMessage("[ERROR] 주문 항목 목록에 메뉴가 누락된 주문 항목이 존재합니다.");
-    }
-
-    @Test
-    @DisplayName("주문 테이블이 비어있는 상태면 예외가 발생한다")
-    void throws_OrderTableIsEmpty() {
-        // given
-        final OrderTable emptyOrderTable = new OrderTable(ORDER_TABLE1_NUMBER_OF_GUESTS, true);
-
-        // when & then
-        assertThatThrownBy(() -> OrderValidator.validateOrderTable(emptyOrderTable))
-                .isInstanceOf(OrderException.CannotOrderStateByOrderTableEmptyException.class)
-                .hasMessage("[ERROR] 주문 테이블이 비어있는 상태일 때 주문할 수 없습니다.");
     }
 }
